requirements.txt
flask==2.3.3
python-dotenv==1.0.0
nltk==3.8.1
pydantic==2.4.2
requests==2.31.0 
instructor==1.7.9
